more wiring, and set frame size
